website/requirements.txt
Flask
Flask-SQLAlchemy
flask-login
plotly==5.10.0